pyproject.toml
[tool.poetry]
name = "fi-simulate"
version = "0.1.0"
description = "Voice AI agent simulator and evaluation harness for LiveKit"
authors = ["Future AGI <no-reply@futureagi.com>"]
readme = "README.md"
# Keep current package path; update this if/when you rename the folder
packages = [
  { include = "fi" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
livekit-agents = { version = "~=1.2", extras = ["openai", "silero"] }
livekit = "*"
livekit-api = "*"
pydantic = ">=2.0.0"
python-dotenv = "*"
pandas = "*"
numpy = "*"
ai-evaluation = ">=0.1.9"
openai = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
black = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


